refactor: MVC 구조에 사용하지 않는 타입삭제 및 이름 변경
