+++ WeatherForecast/WeatherForecast/Models/Decoding/WeatherInfoDecode.swift
@@ -26,6 +26,33 @@ struct Parser {
     }
 }
 
+struct FiveDaysForecastData: Decodable {
+    var list: [ForcastInfomation]
+}
+
+struct ForcastInfomation: Decodable {
+    var date: Int
+    var main: Temperature
+    var weather: [WeatherDetail]
+    
+    enum CodingKeys: String, CodingKey {
+        case date = "dt"
+        case main, weather
+    }
+    
+    struct Temperature: Decodable {
+        var temperature: Double
+        
+        enum CodingKeys: String, CodingKey {
+            case temperature = "temp"
+        }
+    }
+    
+    struct WeatherDetail: Decodable {
+        var icon: String
+    }
+}
+
 struct CurrentWeather: Decodable {
     var coordination: Coordinate
     var weather: [Weather]
@@ -60,9 +87,3 @@ struct CurrentWeather: Decodable {
         }
     }
 }
-
-
-
-
-
-
