--- WeatherForecast/WeatherForecast/Observer/Boxing.swift
@@ -1,28 +1,0 @@
-//
-//  Boxing.swift
-//  WeatherForecast
-//
-//  Created by Do Yi Lee on 2021/10/12.
-//
-
-import Foundation
-
-final class Box<T> {
-    typealias Listener = (T) -> Void
-    var listener: Listener?
-    
-    var value: T {
-        didSet {
-            listener?(value)
-        }
-    }
-    
-    init(_ value: T) {
-        self.value = value
-    }
-    
-    func bind(listener: Listener?) {
-        self.listener = listener
-        listener?(value)
-    }
-}
