+++ WeatherForecast/WeatherForecast/Service/NetworkLayer/Networking/Router/EndPoint.swift
@@ -11,17 +11,12 @@ protocol EndPoint {
     var httpTask: HTTPTask { get }
     var httpMethod: HTTPMethod { get }
     var baseUrl: URL { get }
-    var path: Path { get }
+    var path: URLPath { get }
 }
 
 struct OpenWeatherApi: EndPoint {
     var httpTask: HTTPTask
     var httpMethod: HTTPMethod
     var baseUrl: URL
-    var path: Path
-}
-
-enum Path: String {
-    case weather
-    case forecast
+    var path: URLPath
 }
