+++ WeatherForecast/WeatherForecast/Service/NetworkLayer/Networking/Router/NetworkRouter.swift
@@ -10,59 +10,23 @@ import Foundation
 protocol NetworkRouter {
     associatedtype EndPointType = EndPoint
     
-    func request(_ route: EndPointType, _ session: URLSession, _ completionHandler: @escaping (Data) -> ())
+    func request(_ route: EndPointType, _ session: URLSession)
     func cancel()
 }
 
 final class Router<EndPointType: EndPoint>: NetworkRouter {
     private var task: URLSessionDataTask?
     
-    func request(_ route: EndPointType, _ session: URLSession, _ completionHandler: @escaping (Data) -> ()) {
+    func request(_ route: EndPointType, _ session: URLSession) {
         let request = self.buildRequest(from: route)
         
-        task = session.dataTask(with: request, completionHandler: { [weak self] data, response, error in
-            guard let `self` = self else {
-                return
-            }
-            
-            guard let response = response as? HTTPURLResponse, self.checkStatusCode(response: response) else {
-                return
-            }
-            
-            guard let data = data else {
-                return
-            }
-            
-            completionHandler(data)
-        })
+        task = session.dataTask(with: request)
         
         self.task?.resume()
     }
     
     func cancel() {
         self.task?.cancel()
-    }
-    
-    private func checkStatusCode(response: HTTPURLResponse) -> Bool {
-        switch response.statusCode {
-        case 200:
-            return true
-        case 400:
-            print("잘못된 요청입니다.")
-            return false
-        case 401:
-            print("인증이 잘못되었습니다.")
-            return false
-        case 403:
-            print("해당 정보에 접근할 수 없습니다. ")
-            return false
-        case 500...:
-            print("서버에러")
-            return false
-        default:
-            print("네트워크 요청을 확인 해 보세요")
-            return false
-        }
     }
     
     private func buildRequest(from route: EndPointType) -> URLRequest {
