+++ WeatherForecast/WeatherForecast/Service/NetworkLayer/Networking/Router/URLPath.swift
@@ -1,0 +1,13 @@
+//
+//  HTTPPath.swift
+//  WeatherForecast
+//
+//  Created by Do Yi Lee on 2021/10/13.
+//
+
+import Foundation
+
+enum URLPath: String {
+    case weather
+    case forecast
+}
